feat: remove dependent fetches and make them in parallel
