+++ libs/rest/src/hooks/use-candles.ts
@@ -1,45 +1,10 @@
-import {
-  type QueryClient,
-  queryOptions,
-  useQuery,
-  useQueryClient,
-} from '@tanstack/react-query';
-import {
-  queryKeys as candleDataQueryKeys,
-  retrieveCandleData,
-} from '../queries/candle-data';
+import { useQuery, useQueryClient } from '@tanstack/react-query';
+import { candleDataQueryOptions } from '../queries/candle-data';
 import {
   type Interval,
-  queryKeys as candleIntervalsQueryKeys,
+  candleIntervalQueryOptions,
   getCandleId,
-  retrieveCandleIntervals,
 } from '../queries/candle-intervals';
-import { Time } from '../utils/datetime';
-
-export function candleIntervalQueryOptions(params: { marketId: string }) {
-  return queryOptions({
-    queryKey: candleIntervalsQueryKeys.single(params.marketId),
-    queryFn: () => retrieveCandleIntervals({ marketId: params.marketId }),
-    staleTime: Number.POSITIVE_INFINITY,
-  });
-}
-
-export function candleDataQueryOptions(
-  client: QueryClient,
-  params: {
-    candleId: string;
-    marketId: string;
-    fromTimestamp?: string;
-    toTimestamp?: string;
-  }
-) {
-  return queryOptions({
-    queryKey: candleDataQueryKeys.single(params),
-    queryFn: () => retrieveCandleData(params, client),
-    staleTime: Time.HOUR,
-    enabled: Boolean(params.candleId),
-  });
-}
 
 export function useCandleIntervals(params: { marketId: string }) {
   const queryResult = useQuery(candleIntervalQueryOptions(params));
